refactor(plugin-react): move webpack related logics
