+++ packages/plugin-react/package.json
@@ -24,16 +24,12 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@babel/preset-react": "^7.22.15",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
-    "@rspack/plugin-react-refresh": "0.3.8",
     "@rsbuild/shared": "workspace:*",
-    "babel-plugin-transform-react-remove-prop-types": "0.4.24"
+    "@rspack/plugin-react-refresh": "0.3.8"
   },
   "devDependencies": {
     "@rsbuild/core": "workspace:*",
     "@rsbuild/test-helper": "workspace:*",
-    "@rsbuild/webpack": "workspace:*",
     "@types/node": "^16",
     "typescript": "^5"
   },
